Move duplicate code in a seperate file and tweak tpv values
